[#906] Organize parse utils and formatters
Reorganizes parsing and formatting utilities into dedicated modules for
better maintainability and separation of concerns:

- Splits date handling into parse and format utilities
- Extracts boolean and number parsing into standalone functions
- Simplifies transform schemas by leveraging dedicated parsing functions
- Improves code organization and removes redundant date implementations
